Update to TS@next and fix resulting type errors
Also simplify and modify type aliases where necessary.
Note that I haven't fixed any implicit any errors. I also kept the
balance between d.ts and jsdoc about the same.

In lib/formats.ts, stricter call is unable to handle the multiple
overloads of String.prototype.replace, so I had to manually type it to
the one I wanted.

In lib/q.d.ts, ObjectCoercible and ObjectIndexable were effectively the
same, so I removed ObjectIndexable. Also, `{}` subsumes primitive, so I
changed it to `object`:

```ts
type ObjectCoercible = object | NonNullishPrimitive;
```

I changed some similar types that were almost right:
`Concatable<object>` needs to be `Concatable<{}>` because of the way
it's used in utility.compact, and `object | Primitive` is the same as
`unknown`, so I changed `Value` to `Concatable<unknown>`.

In lib/utils.js, I had to edit `merge` in a few places. At least a
couple of the edits looked like they were fixing potential bugs, but I
had trouble visualising the execution of `merge` so I can't be sure.

1. When `typeof source !== "object"`, it could still be `"boolean"`, but
booleans can't be used to index into Object.prototype, so I added an
additional guard.

2. `typeof target !== "object"` is always false, so I deleted the branch
that used it; the code inside was a type error because the compiler
thinks it can't be reached. I couldn't convince myself that it was
wrong.

3. Narrowing by element access failed, so I had to create a `const tmp =
target[i]` before checking that `typeof tmp === "object"`.

4. In `compact`, and I had to add a type annotation to `queue` and then
a cast when pushing into it. The former isn't that surprising,
considering that it's supposed to have a complex type that we can't
infer. The latter is a result of checks that prove that the object
structure has a recursive level that needs to be compacted, but the
compiler can't follow the checks and needs to be told explicitly about
the type.

In tsconfig.json, I had to manually set `"target": "esnext"` to avoid
some bogus errors.

Breakdown of the changes:
* 4 cases where the compiler failed and needed a workaround
* 2 possible bugs
* 3 cases where the type system is full of confusing, similar types