Move file-specific types back to JSDoc
Global aliases like Nullish are now global so they don't need to be
imported. I also introduced `type NonPrimitive = object` to work around
the way Typescript treats `object` as `any` in JS.

The move required hardly any changes; in fact, I basically restored the
annotations from commit 856582edfff28053801677a5f8e2de075580e830 with
updates for the fixes in my previous commits.

Introducing NonPrimitive required a few changes since `object` is
stricter than `any`. The most annoying is a cast on

```js
/** @type {any} */([]).concat(leaf);
```

Because with strictNullChecks on, the type of `[]` is `never[]`, which
is technically correct but usually annoying. (strictNullChecks is usually
not a good fit for JS codebases that don't want to adopt Typescript
idioms completely, and `[].concat` is the best example of this.)