[New] `parse`/`stringify`: Pass extra key/value argument to `decoder`
A new fourth argument is now provided when invoking the decoder inside parse,
this argument represents whether the first argument (str) is a key or a value.
This can then be used to decode key and values different within the decode function.

A test has been added for this new behavior.