Pass extra key/value argument to encoder function calls
A new fourth argument is now provided when invoking the encoder inside stringify,
this argument represents whether the first argument (str) is a key or a value.
This can then be used to encode key and values different within the encode function.

A test has been added for this new behaviour