Add support stringifying space, pipe, and tab delimited arrays
We use this library for stringifying query strings in an API client generated from an OpenAPI spec. As part of the OpenAPI specification, developers may specify that an array may be delimited with a `comma`, `space`, `pipe`, or `tab` character, as well as repeated or exploded. This commit adds support for stringifying arrays with the additional delimiter types.